\name{cperm}
\alias{cperm}
\alias{dansari}
\alias{equiscores}
\alias{toltest}
\title{Internal exactRankTests functions}
\description{
 Internal exactRankTests functions
}
\usage{
cperm(escores, m, paired = NULL)
dansari(x, m ,n)
equiscores(scores, m, tol = 0.01, fact = NULL)
toltest(x, scores, m)
}
\details{
  Those functions are not to be called by the user.
}
\keyword{distribution}
